Use .progmem.data instead of .progmem section
Due to a change in the linker scripts supplied with avr-binutils in
version 2.30 (2018-01-27), the bare `.progmem` section name is no longer
recognized - one now has to use a section name matching the `.progmem.*`
glob.  Change all references to the deprecated section name in this
repository.

The change in question can be found here:

  https://sourceware.org/git/?p=binutils-gdb.git;a=commitdiff;h=f726cc1cb69ab0e0303209d08effb10b99007080